Stardew Valley: Move test option presets to their own file (#4349)
