+++ worlds/stardew_valley/test/TestFill.py
@@ -1,5 +1,6 @@
-from . import SVTestBase, minimal_locations_maximal_items
+from . import SVTestBase
 from .assertion import WorldAssertMixin
+from .options.presets import minimal_locations_maximal_items
 from .. import options
 from ..mods.mod_data import ModNames
