+++ worlds/stardew_valley/test/TestLogic.py
@@ -3,7 +3,9 @@ import unittest
 from unittest import TestCase, SkipTest
 
 from BaseClasses import MultiWorld
-from . import RuleAssertMixin, setup_solo_multiworld, allsanity_mods_6_x_x, minimal_locations_maximal_items
+from . import setup_solo_multiworld
+from .assertion import RuleAssertMixin
+from .options.presets import allsanity_mods_6_x_x, minimal_locations_maximal_items
 from .. import StardewValleyWorld
 from ..data.bundle_data import all_bundle_items_except_money
 from ..logic.logic import StardewLogic
