+++ worlds/stardew_valley/test/mods/TestMods.py
@@ -1,8 +1,10 @@
 import random
 
 from BaseClasses import get_seed, ItemClassification
-from .. import SVTestBase, SVTestCase, allsanity_mods_6_x_x, fill_dataclass_with_default
+from .. import SVTestBase, SVTestCase
 from ..assertion import ModAssertMixin, WorldAssertMixin
+from ..options.presets import allsanity_mods_6_x_x
+from ..options.utils import fill_dataclass_with_default
 from ... import options, items, Group, create_content
 from ...mods.mod_data import ModNames
 from ...options import SkillProgression, Walnutsanity
