+++ worlds/stardew_valley/test/rules/TestStateRules.py
@@ -1,4 +1,5 @@
-from .. import SVTestBase, allsanity_mods_6_x_x
+from .. import SVTestBase
+from ..options.presets import allsanity_mods_6_x_x
 from ...stardew_rule import HasProgressionPercent
